modifies files due to move of decadevs actions and reducers
