+++ src/compositions/authentication/index.js
@@ -1,7 +1,7 @@
-import React, { Component } from "react";
-import { Box, Grommet } from "grommet";
-import { grommet } from "grommet/themes";
-import Authentication from "../../components/authentication";
+import React, { Component } from 'react';
+import { Box, Grommet } from 'grommet';
+import { grommet } from 'grommet/themes';
+import Authentication from '../../components/authentication/';
 
 class App extends Component {
   render() {
@@ -12,8 +12,8 @@ class App extends Component {
           <Box
             style={{
               paddingBottom: `48px`,
-              borderRadius: "10px",
-              boxShadow: "0px 0px 20px 5px rgba(0, 0, 0, 0.1)"
+              borderRadius: '10px',
+              boxShadow: '0px 0px 20px 5px rgba(0, 0, 0, 0.1)'
             }}
             width="medium"
             pad="medium"
